GCC: fix some warnings mostly related to C++
This commit prepares for a GCC upgrade; the build is performed with `./configure --enable-stop-on-warning CC=gcc-16 CXX=g++-16 F77=gfortran-16`.

See https://gitlab.com/scilab/scilab/-/merge_requests/1774

+++ scilab/modules/scicos/src/cpp/SSPResource_save.cpp
@@ -348,18 +348,18 @@ SSPResource::Result SSPResource::save(const char* uri)
 static bool is_empty_matrix(const std::vector<double>& v)
 {
     // v == {1, 2, 0, 0, 0}
-    return v.size() == 5 && v[0] == sci_matrix && v[1] == 2. && v[2] == 0. && v[3] == 0. && v[4] == 0.;
+    return v.size() == 5 && static_cast<int>(v[0]) == sci_matrix && v[1] == 2. && v[2] == 0. && v[3] == 0. && v[4] == 0.;
 }
 
 static bool is_empty_list(const std::vector<double>& v)
 {
     // v == {15, 0}
-    return v.size() == 2 && v[0] == sci_list && v[1] == 0.;
+    return v.size() == 2 && static_cast<int>(v[0]) == sci_list && v[1] == 0.;
 }
 
 static bool is_string_vector(const std::vector<double>& v)
 {
-    return v.size() > 2 && v[0] == sci_strings && v[1] != 0;
+    return v.size() > 2 && static_cast<int>(v[0]) == sci_strings && v[1] != 0;
 }
 
 /* helper function to decode simple string EXPRS */
